third_party/WebKit/LayoutTests/fast/events/drag-and-drop-autoscroll-use-count.html

<!DOCTYPE html>
<style type="text/css">
#draggable {
    padding: 5pt;
    border: 3px solid #00cc00;
    background: #00cccc;
    width: 80px;
    cursor: pointer;
}
#scrollable {
    height: 200px;
    overflow: auto;
    border: solid 3px #cc0000;
    font-size: 80px;
}
</style>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="draggable" draggable="true">
<div id="scrollable">
    <p>a</p>
    <p>b</p>
    <p>c</p>
    <p>d</p>
    <p>e</p>
    <p>f</p>
    <p>g</p>
    <p>h</p>
    <p>i</p>
    <p>j</p>
</div>
<script>
test(function() {
    if (!window.eventSender)
        return;
    var DragAndDropStart = 1554;
    var scrollable = document.getElementById("scrollable");
    var draggable = document.getElementById("draggable");
    var dragStartX = draggable.offsetLeft + 5;
    var dragStartY = draggable.offsetTop + 5;
    var dragEndX = scrollable.offsetLeft + 5;
    var dragEndY = scrollable.offsetTop + scrollable.offsetHeight - 10;

    eventSender.mouseMoveTo(dragStartX, dragStartY);
    eventSender.mouseDown();
    eventSender.mouseMoveTo(dragEndX, dragEndY);
    eventSender.mouseUp();
    assert_true(internals.isUseCounted(document, DragAndDropStart));
}, "Tests that selection-drag autoscroll is recorded in UMA usecounter.");  
</script>
